+++ CHANGELOG.md
@@ -1,7 +1,10 @@
+# v0.1.2
+- add `String#from_albhed` method
+
 # v0.1.1
 Edit CHANGELOG.md
 
 # v0.1.0
 First Release
 
-- add `to_albhed` method
+- add `String#to_albhed` method